refactor: collapse logging package into __init__ for cleaner import
Move logger.py contents into the logging package __init__.py so call
sites import from inference_perf.observability.logging directly instead
of the redundant ...logging.logger path.
